+++ src/xenia/vfs/premake5.lua
@@ -39,4 +39,5 @@ project("xenia-vfs-dump")
   resincludedirs({
     project_root,
   })
+include("testing")
