+++ src/xenia/vfs/testing/premake5.lua
@@ -1,0 +1,8 @@
+project_root = "../../../.."
+include(project_root.."/tools/build")
+
+test_suite("xenia-vfs-tests", project_root, ".", {
+  links = {
+    "xenia-vfs",
+  },
+})
